Supporting url for linkmove -> extractorsDir
